% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.r
\name{calibration_fusion}
\alias{calibration_fusion}
\title{Fuses multiple LR systems and provides single calibrated output}
\usage{
calibration_fusion(test_list, dev_list = NULL, log = TRUE)
}
\arguments{
\item{test_list}{A list of LR matrices to be calibrated and fused. The input must be a LIST. Any data frame or matrix will work, as long as it follows the same format as the output in MVKD_loop (a row for each suspect, a column for each offender, LR value not transformed in any way). The LR matrices should all be of the same dimensions.}

\item{dev_list}{(Optional) A list of LR matrices to be trained on. The list must be of the same length as test_list and matrices within must correspond with those in test_list. If left blank, training will be based on test_list.}

\item{log}{Boolean. LRs already log-transformed (TRUE; default) or not (FALSE)?}
}
\value{
A named list following the exact format as MVKD_loop: likelihood_ratio_matrix, cllr and eer.
}
\description{
Combines LR scores from (multiple) test systems to provide a single set of fused LR scores, based on a logistric-regression model trained with the training data input. Self-calibration based on training with the same set of data is possible, in which case the training data argument should be omitted.
An R implementation of \code{calibration_fusion.m} by Hughes (2015) and \code{train_llr_fusion_robust.m} by Morrison (2009) (http://geoff-morrison.net/#TrainFus).
}
