% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{cllr}
\alias{cllr}
\title{Calculates the log likelihood ratio cost}
\usage{
cllr(LLR_SS, LLR_DS, log = TRUE)
}
\arguments{
\item{LLR_SS}{A vector of LRs from same-speaker comparisons.}

\item{LLR_DS}{A vector of LRs from different-speaker comparisons.}

\item{log}{(optional) Boolean. Specifies whether LRs supplied are already in (natural) log form (default TRUE).}
}
\value{
A single numerical Cllr.
}
\description{
Calculates the log likelihood ratio cost
}
