% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{EER_linear}
\alias{EER_linear}
\title{Calculates the equal error rate}
\usage{
EER_linear(LR_SS, LR_DS, log = FALSE)
}
\arguments{
\item{log}{(optional) Boolean. Specifies whether LRs supplied are already in (natural) log form (default FALSE).}

\item{LLR_SS}{A vector of LRs from same-speaker comparisons.}

\item{LLR_DS}{A vector of LRs from different-speaker comparisons.}
}
\value{
A single numerical EER between 0 and 1 (inclusive).
}
\description{
Calculates the equal error rate
}
