% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM_UBM.r
\name{GMM_UBM_loop}
\alias{GMM_UBM_loop}
\title{Performs GMM-UBM-based LR calculation on a set of speakers}
\usage{
GMM_UBM_loop(
  bg_data,
  by_speaker_data,
  test_speakers,
  G = 8,
  r = 16,
  background_model = NULL
)
}
\arguments{
\item{bg_data}{A data frame of background data. The first column identifies speakers. All other columns contain data.}

\item{by_speaker_data}{A named list of 2 sub-lists: \code{suspect_data} and \code{offender_data}. Contains data in the \code{test_speakers} set, divided by speaker. Each is a named list of data frames, with speaker IDs as the names.}

\item{test_speakers}{A vector of test speakers.}

\item{G}{Number of components in the Gaussian Mixture Model.}

\item{r}{Relevance factor for speaker adaptation.}

\item{background_model}{(Optional) Model pre-fitted on bg_data. Useful when the same model is used for different test data sets.}
}
\value{
A named list of 3 items:
\itemize{
\item \code{likelihood_ratio_matrix}: A data frame. Rows and columns are named after the speaker identifiers. Each row and column represents a speaker as suspect and offender respectively, and each cell contains a single logLR score.
\item \code{cllr}: Numeric. Reports the logLR cost.
\item \code{eer}: Numeric. Reports the equal error rate (between 0 and 1).
}
}
\description{
Models reference population using a single GMM (UBM) and models suspect by adapting from UBM. Called by LR_test when the mode GMM-UBM is chosen. Procedure is based on \code{Reynolds, D.A., Quatieri, T.F. & Dunn, R.B. (2000). Speaker Verification Using Gaussian Mixture Models. Digital Signal Processing, 10(1-3), 19-41.}
}
