% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar_kernel_LR.r
\name{multivar_kernel_LR}
\alias{multivar_kernel_LR}
\title{Calculates likelihood ratio based on MVKD}
\usage{
multivar_kernel_LR(sus_data, off_data, bg, bg_stat = NULL)
}
\arguments{
\item{sus_data}{A matrix containing all suspect data to be compared}

\item{off_data}{A matrix containing all offender data to be compared}

\item{bg}{A data frame that contains all data in the background model. The first column identifies the speaker.}

\item{bg_stat}{(Optional) Data frame of pre-calculated means and covariance matrices. Avoid providing your own bg_stat at all cost, because it will be generated within if left blank. It exists primarily for use in conjunction with MVKD_loop(), which will generate it in prior to save computation time.}
}
\value{
Numeric. The log-likelihood ratio.
}
\description{
This function is in part an adaptation of \code{Morrison, G.S. (2007). Matlab implementation of Aitken & Lucy\'s (2004) forensic likelihood-ratio software using multivariate-kernel-density estimation. [Software]}. (http://geoff-morrison.net/#MVKD)
For details of how MVKD works, see \code{Aitken, C.G.G. & Lucy, D. (2004). Evaluation of trace evidence in the form of multivariate data. Applied Statistics, 54, 109-122.}
}
