% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVKD.r
\name{MVKD}
\alias{MVKD}
\title{Performs MVKD-based LR calculation on a set of speakers}
\usage{
MVKD(bg_data, by_speaker_data, test_speakers, bg_speakers)
}
\arguments{
\item{bg_data}{A data frame. The first column identifies speakers.}

\item{by_speaker_data}{A named list of 2 sub-lists: \code{suspect_data} and \code{offender_data}. Contains data in the \code{test_speakers} set, divided by speaker. Each is a named list of data frames, with speaker IDs as the names.}

\item{test_speakers}{(optional) A vector. Specifies the set of speakers for which LR calculations should be derived. Takes all speakers from \code{data} if unspecified.}

\item{bg_speakers}{(optional) A vector. Specifies the set of speakers to form the background model for LR calculations. Equal to \code{test_speakers} if unspecified.}
}
\value{
A named list of 3 items:
\itemize{
\item \code{likelihood_ratio_matrix}: A data frame. Rows and columns are named after the speaker identifiers. Each row and column represents a speaker as suspect and offender respectively, and each cell contains a single LR score.
\item \code{cllr}: Numeric. Reports the logLR cost.
\item \code{eer}: Numeric. Reports the equal error rate (between 0 and 1).
}
}
\description{
Models reference population using a multivariate kernel density and models suspect by normal density
}
