% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVKD_loop.r
\name{MVKD_loop}
\alias{MVKD_loop}
\title{Performs LR calculation on a set of speakers}
\usage{
MVKD_loop(
  data,
  test_speakers = NULL,
  bg_speakers = NULL,
  data_col = NULL,
  test_data = NULL,
  test_data_col = NULL,
  cross_full = TRUE
)
}
\arguments{
\item{data}{A data frame. The first column identifies speakers.}

\item{test_speakers}{(optional) A vector. Specifies the set of speakers for which LR calculations should be derived. Takes all speakers from \code{data} if unspecified.}

\item{bg_speakers}{(optional) A vector. Specifies the set of speakers to form the background model for LR calculations. Equal to \code{test_speakers} if unspecified.}

\item{data_col}{(optional) A vector. Specifies which columns in \code{data} to take into account in LR calculation. Uses all available columns if unspecified (excludes first column, which identifies speakers).}

\item{test_data}{(optional) A data frame in the same format as \code{data}. If specified, offender/disputed-speaker data will be taken from here. Suspect/known-speaker data and background data will be taken from \code{data}.}

\item{test_data_col}{(optional) A vector. Specifies which columns in \code{data} to take into account in LR calculation. Follows \code{data_col} if unspecified.}

\item{cross_full}{(optional) Boolean. When \code{test_data} and \code{test_data_col} are specified, determines if the full set of data from each speaker are used in the comparison. Default TRUE. If FALSE, suspect data come from first half of \code{data} and questioned data come from second half of \code{test_data}, to simulate the scenario when \code{test_data} is not specified. Useful for keeping number of data points constant, in the case of comparisons involving non-contemporaneous recordings or mismatched conditions and the results of which are compared with contemporaneous comparisons.}
}
\value{
A named list of 3 items:
\itemize{
\item \code{likelihood_ratio_matrix}: A data frame. Rows and columns are named after the speaker identifiers. Each row and column represents a speaker as suspect and offender respectively, and each cell contains a single LR score.
\item \code{cllr}: Numeric. Reports the logLR cost.
\item \code{eer}: Numeric. Reports the equal error rate (between 0 and 1).
}
}
\description{
Iterates through all pairings in a set of test speakers and calculates the likelihood ratio for each speaker-pair
}
