% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_functions.r
\name{zooify}
\alias{zooify}
\title{Calculate SS and DS mean scores from likelihood ratio matrix}
\usage{
zooify(LR.matrix)
}
\arguments{
\item{LR.matrix}{Data frame of likelihood ratio scores. Row and column names must correspond to the speaker names/identifiers. LR scores should not be transformed in any way (e.g. logarithmically).}
}
\value{
Data frame where each row contains the SS and DS mean scores for a speaker.
}
\description{
Not ready for use
}
