% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_functions.r
\name{zooplot}
\alias{zooplot}
\title{Zooplot}
\usage{
zooplot(LR.zoo, quantile = TRUE)
}
\arguments{
\item{LR.zoo}{Zoo data frame returned by \code{zooify}.}

\item{quantile}{Plot line segments marking quantiles of scores if TRUE.}
}
\description{
Zooplot
}
